requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.11
packaging==25.0
pillow==12.1.0
requests==2.32.5
urllib3==2.6.3
